requirements.txt
# API server
fastapi
uvicorn[standard]
python-multipart

# Data collection
requests
yfinance
pandas
ta

# Env
python-dotenv
